+++ packages/account/src/account.ts
@@ -53,8 +53,8 @@ import {
 import { mergeQuantities } from './providers/utils/merge-quantities';
 import { serializeProviderCache } from './providers/utils/serialization';
 import { AbstractAccount } from './types';
-import { splitCoinsIntoBatches } from './utils/consolidate-coins';
 import { assembleTransferToContractScript } from './utils/formatTransferToContractScriptData';
+import { splitCoinsIntoBatches } from './utils/split-coins-into-batches';
 
 export type TxParamsType = Pick<
   ScriptTransactionRequestLike,
